+++ src/app.module.ts
@@ -13,6 +13,7 @@ import { HealthModule } from "./modules/health/health.module";
 import { MediaModule } from "./modules/media/media.module";
 import { MetadataModule } from "./modules/metadata/metadata.module";
 import { ProgressModule } from "./modules/progresses/progress.module";
+import { SavefileModule } from "./modules/savefiles/savefile.module";
 import { UsersModule } from "./modules/users/users.module";
 
 @Module({
@@ -23,6 +24,7 @@ import { UsersModule } from "./modules/users/users.module";
     GamesModule,
     UsersModule,
     ProgressModule,
+    SavefileModule,
     MetadataModule,
     AdminModule,
     ScheduleModule.forRoot(),