+++ src/modules/progresses/progress.service.ts
@@ -89,7 +89,7 @@ export class ProgressService implements OnApplicationBootstrap {
     }
   }
 
-  public async delete(
+  public async deleteByProgressId(
     progressId: number,
     executorUsername: string,
   ): Promise<Progress> {
@@ -103,6 +103,28 @@ export class ProgressService implements OnApplicationBootstrap {
       executorUsername,
     );
 
+    const softRemoveResult = await this.progressRepository.softRemove(progress);
+    this.logger.log({
+      message: `Soft-deleted progress.`,
+      progress,
+    });
+    return softRemoveResult;
+  }
+
+  public async deleteByUserIdAndGameId(
+    userId: number,
+    gameId: number,
+    executorUsername: string,
+  ): Promise<Progress> {
+    await this.usersService.checkIfUsernameMatchesIdOrIsAdmin(
+      userId,
+      executorUsername,
+    );
+    const progress = await this.findOneByUserIdAndGameIdOrReturnEmptyProgress(
+      userId,
+      gameId,
+      { loadDeletedEntities: true, loadRelations: true },
+    );
     const softRemoveResult = await this.progressRepository.softRemove(progress);
     this.logger.log({
       message: `Soft-deleted progress.`,